Fixed alot of bugs in ads-stack.  Added some changes to the default maadi, but will need to convert the default directory for bsh back.
Added some stuff so that the AlebgraicADSStack can call it's super components (push, pop, etc) without error.

This is MOSTLY working.  However, there is an issue with calling NEWSTACKINDEX and NEWSTACKSIZE.  The logger is telling me that newStackIndex and newStackSize are not supported procedures.  This might be due to the failsafe within the NULCONSTRUCT method, but I am unable to locate where this error is being thrown.

SFH.