Adding JavaQueue.rb.
This is almost the same as JavaStack, but I made it it's own class to handle specific differences for issues versus api on Stack versus Queue.

Example:
atIndex->get
push->add
pop->remove

However, the JavaQueue will still work with the ADS stack experts.

Still need to do this for ruby and test both implementations to see if they work.

SFH