Looks like I figured out a few things about ruby shoes.
I managed to get some kind of UI in the words, but it is sloppy.  And does not use stack and flows very well (makes a list of KV pairs).

Also, I cant seem to pass attributes to shoe objects from other classes.  I have no idea how I am going to get this to work in the long run.

Added Prompt.rb which contains most of the experimental ui work.

SFH.