Added wrapper classes for XMLElement and attribute.
This works sorta right now, but this framework is more extensible and will allow actions like editing xml and loading xml to work wonders.

Still working out the kinks, need to finish XMLElementWrapper so that I can start on XMLAttribute as a widget, but for now this is much better than before.

SFH